+++ lib/src/dart_binaries.dart
@@ -17,23 +17,31 @@ class DartBinaries {
 
   static const DART2JS_BIN_NAME = 'dart2js';
 
+  static const DART_BIN_NAME = 'dart';
+
   /// Path to the Content Shell executable.
   String contentShellBin;
 
   /// Path to the pub executable.
   String pubBin;
 
+  /// Path to the dart executable.
+  String dartBin;
+
   /// Path to the dart2js executable.
   String dart2jsBin;
 
-  DartBinaries(this.contentShellBin, this.pubBin, this.dart2jsBin);
+  DartBinaries(this.contentShellBin, this.pubBin, this.dart2jsBin,
+               this.dartBin);
 
   factory DartBinaries.withDefaults() =>
-      new DartBinaries(CONTEST_SHELL_BIN_NAME, PUB_BIN_NAME, DART2JS_BIN_NAME);
+      new DartBinaries(CONTEST_SHELL_BIN_NAME, PUB_BIN_NAME, DART2JS_BIN_NAME,
+                       DART_BIN_NAME);
 
   /// Checks that all the Dart SDK binaries are accessible and working.
   /// If some binaries are not in the PATH a [ArgumentError] will be thrown.
   void checkDartSdkBinaries() {
+    dartBin = _checkBinary(dartBin, "--dart-bin", "dart", "Dart");
     pubBin = _checkBinary(pubBin, "--pub-bin", "pub", "Pub");
     dart2jsBin =
         _checkBinary(dart2jsBin, "--dart2js-bin", "dart2js", "dart2js");